mtc-campus-system/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
Pillow==10.3.0
qrcode[pil]==7.4.2
celery==5.4.0
redis==5.0.4
python-dotenv==1.0.1
django-cors-headers==4.3.1
django-filter==24.2
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
setuptools>=69.0.0
